+++ src/examples/tokens/SimpleTokenExample.sol
@@ -11,7 +11,7 @@ contract SimpleTokenExample is ERC20, CsrRewardsERC20 {
         string memory _name, 
         string memory _symbol,
         bool _usingFee,
-        uint8 _feeBasisPoints,
+        uint16 _feeBasisPoints,
         uint _supply
     ) ERC20(_name, _symbol) CsrRewardsERC20(_usingFee, _feeBasisPoints) {
         _mint(msg.sender, _supply);
