+++ src/examples/tokens/TaxTokenExample.sol
@@ -19,7 +19,7 @@ contract TaxTokenExample is ERC20, CsrRewardsERC20 {
         string memory _name, 
         string memory _symbol,
         bool _usingFee,
-        uint8 _feeBasisPoints,
+        uint16 _feeBasisPoints,
         uint _supply,
         TaxCollector[] memory _taxCollectors
     ) ERC20(_name, _symbol) CsrRewardsERC20(_usingFee, _feeBasisPoints) {
